Fixing Bugs and Errors
- Use instance DeleteUser in UsersControllers for better encapsulation
- Handle Status, TransactionType as nullable in DTO mapping
- Fix QRCodeID assignment in _AddQRCode; update mode after add
- Remove static Confirm, add instance Delete in Transactions
- Read Status as Int32 in QRsData to match schema changes

+++ BreadApp_DL/QRsData.cs
@@ -92,7 +92,7 @@ namespace BreadApp_DL
                 reader.IsDBNull(reader.GetOrdinal("UserID")) ? null : reader.GetInt32(reader.GetOrdinal("UserID")),
                 reader.IsDBNull(reader.GetOrdinal("BreadPointID")) ? null : reader.GetInt32(reader.GetOrdinal("BreadPointID")),
                 reader.IsDBNull(reader.GetOrdinal("PortionCount")) ? null : reader.GetInt32(reader.GetOrdinal("PortionCount")),
-                reader.IsDBNull(reader.GetOrdinal("Status")) ? null : reader.GetInt16(reader.GetOrdinal("Status")),
+                reader.IsDBNull(reader.GetOrdinal("Status")) ? null : reader.GetInt32(reader.GetOrdinal("Status")),
                 reader.IsDBNull(reader.GetOrdinal("Token")) ? null : reader.GetString(reader.GetOrdinal("Token")),
                 reader.IsDBNull(reader.GetOrdinal("CreatedAt")) ? null : reader.GetDateTime(reader.GetOrdinal("CreatedAt")),
                 reader.IsDBNull(reader.GetOrdinal("ExpiresAt")) ? null : reader.GetDateTime(reader.GetOrdinal("ExpiresAt")),
